+++ Source/DotNetWorkQueue/Metrics/Decorator/IClearExpiredMessagesDecorator.cs
@@ -37,8 +37,8 @@ namespace DotNetWorkQueue.Metrics.Decorator
             IConnectionInformation connectionInformation)
         {
             var name = "ClearExpiredMessages";
-            _timer = metrics.Timer($"{connectionInformation.QueueName}.{name}.ClearMessages.ResetTimer", Units.Calls);
-            _counter = metrics.Counter($"{connectionInformation.QueueName}.{name}.ClearMessages.ResetCounter", Units.Items);
+            _timer = metrics.Timer($"dotnetworkqueue.{connectionInformation.QueueName}.{name}.ClearMessages.ResetTimer", Units.Calls);
+            _counter = metrics.Counter($"dotnetworkqueue.{connectionInformation.QueueName}.{name}.ClearMessages.ResetCounter", Units.Items);
             _handler = handler;
         }